Add SpatialIndex trait
