+++ rust/sedona-spatial-join/src/index.rs
@@ -16,17 +16,19 @@
 // under the License.
 
 pub(crate) mod build_side_collector;
+pub(crate) mod default_spatial_index;
+pub(crate) mod default_spatial_index_builder;
 mod knn_adapter;
 pub(crate) mod memory_plan;
 pub(crate) mod partitioned_index_provider;
 pub(crate) mod spatial_index;
-pub(crate) mod spatial_index_builder;
 
 pub(crate) use build_side_collector::{
     BuildPartition, BuildSideBatchesCollector, CollectBuildSideMetrics,
 };
-pub use spatial_index::SpatialIndex;
-pub use spatial_index_builder::{SpatialIndexBuilder, SpatialJoinBuildMetrics};
+pub(crate) use spatial_index::SpatialIndex;
+
+pub use default_spatial_index_builder::{SpatialIndexBuilder, SpatialJoinBuildMetrics};
 use wkb::reader::Wkb;
 
 /// The result of a spatial index query
