+++ rust/sedona-spatial-join/src/index/partitioned_index_provider.rs
@@ -28,11 +28,12 @@ use std::sync::Arc;
 use tokio::sync::mpsc;
 
 use crate::evaluated_batch::evaluated_batch_stream::external::ExternalEvaluatedBatchStream;
+use crate::index::spatial_index::SpatialIndexRef;
 use crate::index::BuildPartition;
 use crate::partitioning::stream_repartitioner::{SpilledPartition, SpilledPartitions};
 use crate::utils::disposable_async_cell::DisposableAsyncCell;
 use crate::{
-    index::{SpatialIndex, SpatialIndexBuilder, SpatialJoinBuildMetrics},
+    index::{SpatialIndexBuilder, SpatialJoinBuildMetrics},
     partitioning::SpatialPartition,
     spatial_predicate::SpatialPredicate,
 };
@@ -49,7 +50,7 @@ pub(crate) struct PartitionedIndexProvider {
     data: BuildSideData,
 
     /// Async cells for indexes, one per regular partition
-    index_cells: Vec<DisposableAsyncCell<SharedResult<Arc<SpatialIndex>>>>,
+    index_cells: Vec<DisposableAsyncCell<SharedResult<SpatialIndexRef>>>,
 
     /// The memory reserved in the build side collection phase. We'll hold them until
     /// we don't need to build spatial indexes.
@@ -145,7 +146,7 @@ impl PartitionedIndexProvider {
     pub async fn build_or_wait_for_index(
         &self,
         partition_id: u32,
-    ) -> Option<Result<Arc<SpatialIndex>>> {
+    ) -> Option<Result<SpatialIndexRef>> {
         let cell = match self.index_cells.get(partition_id as usize) {
             Some(cell) => cell,
             None => {
@@ -194,7 +195,7 @@ impl PartitionedIndexProvider {
         }
     }
 
-    async fn maybe_build_index(&self, partition_id: u32) -> Option<Result<Arc<SpatialIndex>>> {
+    async fn maybe_build_index(&self, partition_id: u32) -> Option<Result<SpatialIndexRef>> {
         match &self.data {
             BuildSideData::SinglePartition(build_partition_opt) => {
                 if partition_id != 0 {
@@ -237,7 +238,7 @@ impl PartitionedIndexProvider {
         }
     }
 
-    pub async fn wait_for_index(&self, partition_id: u32) -> Option<Result<Arc<SpatialIndex>>> {
+    pub async fn wait_for_index(&self, partition_id: u32) -> Option<Result<SpatialIndexRef>> {
         let cell = match self.index_cells.get(partition_id as usize) {
             Some(cell) => cell,
             None => {
@@ -268,7 +269,7 @@ impl PartitionedIndexProvider {
     async fn build_index_for_single_partition(
         &self,
         build_partitions: Vec<BuildPartition>,
-    ) -> Result<Arc<SpatialIndex>> {
+    ) -> Result<SpatialIndexRef> {
         let mut index_builder = SpatialIndexBuilder::new(
             Arc::clone(&self.schema),
             self.spatial_predicate.clone(),
@@ -284,14 +285,13 @@ impl PartitionedIndexProvider {
             index_builder.add_stream(stream, geo_statistics).await?;
         }
 
-        let index = index_builder.finish()?;
-        Ok(Arc::new(index))
+        index_builder.finish()
     }
 
     async fn build_index_for_spilled_partition(
         &self,
         spilled_partition: SpilledPartition,
-    ) -> Result<Arc<SpatialIndex>> {
+    ) -> Result<SpatialIndexRef> {
         let mut index_builder = SpatialIndexBuilder::new(
             Arc::clone(&self.schema),
             self.spatial_predicate.clone(),
@@ -345,14 +345,13 @@ impl PartitionedIndexProvider {
 
         index_builder.merge_stats(geo_statistics);
 
-        let index = index_builder.finish()?;
-        Ok(Arc::new(index))
+        index_builder.finish()
     }
 }
 
 async fn get_index_from_cell(
-    cell: &DisposableAsyncCell<SharedResult<Arc<SpatialIndex>>>,
-) -> Option<Result<Arc<SpatialIndex>>> {
+    cell: &DisposableAsyncCell<SharedResult<SpatialIndexRef>>,
+) -> Option<Result<SpatialIndexRef>> {
     match cell.get().await {
         Some(Ok(index)) => Some(Ok(index)),
         Some(Err(shared_err)) => Some(Err(DataFusionError::Shared(shared_err))),
@@ -546,7 +545,7 @@ mod tests {
             .build_or_wait_for_index(0)
             .await
             .expect("partition exists")?;
-        assert_eq!(first_index.indexed_batches.len(), 1);
+        assert_eq!(first_index.num_indexed_batches(), 1);
         assert_eq!(provider.num_loaded_indexes(), 1);
 
         let cached_index = provider
@@ -591,13 +590,13 @@ mod tests {
         let idx_one = idx_one.expect("partition exists")?;
         let idx_two = idx_two.expect("partition exists")?;
         assert!(Arc::ptr_eq(&idx_one, &idx_two));
-        assert_eq!(idx_one.indexed_batches.len(), 1);
+        assert_eq!(idx_one.num_indexed_batches(), 1);
 
         let second_partition = provider
             .build_or_wait_for_index(1)
             .await
             .expect("second partition exists")?;
-        assert_eq!(second_partition.indexed_batches.len(), 1);
+        assert_eq!(second_partition.num_indexed_batches(), 1);
         assert_eq!(provider.num_loaded_indexes(), 2);
         Ok(())
     }
